+++ NEWS.md
@@ -1,5 +1,8 @@
 # xml2 1.5.2
 
+* `xml_add_child()` without `.where` argument is now much faster. The default
+  is now `NULL` and still means "append after the last child".
+
 * `read_html()` now defaults to `encoding = "UTF-8"` to prevent double-encoding
   of UTF-8 content on Windows with codepage 65001 (#490).